fixing value type for kalman_noise
in TF 1.5, the code won't be able to run because the kalman_noise is set to 3e-2 which is a float, but in the flags it's labelled as int, and tensorflow will fail. This PR changes it to it's correct type so it can be run under TF 1.5. 

Other code works fine under TF1.5.