updated working script
In earlier version of the script, at line 115 and 123, the function make_gaussian was called which is not defined in utils.py which was imported earlier. The definition of the function make_gaussian was in cpm_utils.py so I have updated the import. Moreover, the make_gaussian image should be replaced with gaussian_img function defined in cpm_utils.py with appropriate arguments to generate the desired heatmaps.